Post and files CRUD Successfully

+++ src/main/java/egovframework/example/admin/boards/service/BoardsService.java
@@ -9,6 +9,6 @@ public interface BoardsService {
 	List<Map<String, Object>> findByBoard(String boardId, int size, int offset) throws Exception;
 	PostVO getPostDetail(String postId) throws Exception;
 	Long createPost(PostVO dto) throws Exception;
-	int updatePost(PostVO dto) throws Exception;
+	Long updatePost(PostVO dto) throws Exception;
 	int deletePost(Map<String,Object> param) throws Exception;
 }
